+++ packages/api/api/analytics/military_score_inference/policy_ladder_tier_budget.py
@@ -7,6 +7,14 @@ from dataclasses import dataclass
 
 from api.analytics.military_score_inference.inference_cancel import InferenceCancelToken
 from api.analytics.military_score_inference.policy_ladder_state import PolicyLadderState
+from api.analytics.military_score_inference.tier_policy import InferenceTierPolicyStep
+
+__all__ = (
+    "TierStepRun",
+    "ensure_ladder_clock_started",
+    "remaining_time",
+    "tier_step_allowance_seconds",
+)
 
 
 def remaining_time(started_at: float, time_limit_seconds: float | None) -> float:
@@ -22,8 +30,35 @@ def ensure_ladder_clock_started(state: PolicyLadderState, *, now: float | None =
     return state.started_at
 
 
+def tier_step_allowance_seconds(
+    steps: tuple[InferenceTierPolicyStep, ...],
+    step_index: int,
+    *,
+    global_remaining_seconds: float,
+) -> tuple[float, float, float]:
+    """Return ``(allowance, reserved_for_later, spendable)`` for one ladder step.
+
+    Soft-global remainder **steers** the target slice: later steps' ``min_seconds``
+    are reserved so earlier steps prefer not to consume them
+    (``spendable = max(0, global_remaining - reserved)``, then capped by
+    ``max_seconds``). The current step's ``min_seconds`` is an **absolute floor**
+    on allowance even when that exceeds ``spendable`` / soft-global remainder
+    (intentional overshoot so high-prior aggregate tiers still run).
+    """
+    if step_index < 0 or step_index >= len(steps):
+        raise ValueError(f"step_index {step_index} out of range for {len(steps)} steps")
+    step = steps[step_index]
+    reserved = sum(later.min_seconds for later in steps[step_index + 1 :])
+    spendable = max(0.0, float(global_remaining_seconds) - reserved)
+    steered = spendable
+    if step.max_seconds is not None:
+        steered = min(steered, step.max_seconds)
+    allowance = max(step.min_seconds, steered)
+    return allowance, reserved, spendable
+
+
 @dataclass
-class _TierStepRun:
+class TierStepRun:
     """Cancel and time-budget guards shared across one tier step.
 
     Soft-global wall budget **steers** each step's target allowance at dispatch
